==> src/2-aim1/0_run_scripts.R <==


rm(list=ls())
source(here::here("0-config.R"))

source(here::here("src/2-aim1/1-combine-datasets.R"))
source(here::here("src/2-aim1/2-aim-1-analysis.R"))
source(here::here("src/2-aim1/3-aim-1-abund-analysis.R"))
source(here::here("src/2-aim1/4-aim-1-subgroup-analysis.R"))
source(here::here("src/2-aim1/5-pool-results.R"))


source(here::here("src/2-aim1/6-tables.R"))
source(here::here("src/2-aim1/7-abundance-tables.R"))

source(here::here("src/2-aim1/8-figures.R"))
source(here::here("src/2-aim1/9-figures-subgroup_PD.R"))
source(here::here("src/2-aim1/10-aim-1-prescreen-sens-analysis.R"))


#source(here::here("src/2-aim1/10-zoonotic-animals.R"))
#source(here::here("src/2-aim1/11-sensitivity-analysis-study-vs-trial.R"))
